Revert "Splitting up rs files for consistency and readability."
This reverts commit 154d90fe5fd364bbe5ad5f1a28753063604a8b74.
